meta: continue porting the system to i686
This patch does enough work to get the kenrel to compile, but not
link. Additionally, this the patchelf utility is now used to set
libgcc_s.so as a dependency of libc.so. This ensures every program
is which links against libc will get this library, and also lets
libc use libgcc functions, which is very nice for the i686 port.

+++ kernel/net/tcp_socket.c
@@ -219,10 +219,10 @@ int tcp_send_segments(struct socket *socket) {
         uint32_t data_queued = data_available - (tcb->send_next - tcb->send_unacknowledged);
 
         uint32_t data_to_send = 0;
-        if ((MIN(usable_window, data_queued) >= (ssize_t) segment_size) ||
-            ((socket->tcp_nodelay || tcb->send_next == tcb->send_unacknowledged) && push_data && data_queued <= usable_window) ||
+        if ((MIN((uint32_t) usable_window, data_queued) >= segment_size) ||
+            ((socket->tcp_nodelay || tcb->send_next == tcb->send_unacknowledged) && push_data && data_queued <= (uint32_t) usable_window) ||
             ((socket->tcp_nodelay || tcb->send_next == tcb->send_unacknowledged) &&
-             MIN(usable_window, data_queued) >= tcb->send_window_max / 2)) {
+             MIN((uint32_t) usable_window, data_queued) >= tcb->send_window_max / 2)) {
             data_to_send = MIN(MIN(data_queued, segment_size), (size_t) usable_window);
         } else if (push_data && data_queued != 0) {
             // FIXME: Set up a timer to send the data in ~0.5s
